Moved AcrylicPane
